requirements.txt
mariadb==1.1.14
packaging==26.0
python-dotenv